+++ src/applications/personalization/profile/components/hub/Hub.jsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 
 import {
   hasBadAddress as hasBadAddressSelector,
-  selectProfileContactsToggle,
+  selectProfileContactsPageToggle,
 } from '@@profile/selectors';
 import { useSelector } from 'react-redux';
 
@@ -17,7 +17,9 @@ import { HubCard } from './HubCard';
 export const Hub = () => {
   const { label, link } = useSignInServiceProvider();
   const hasBadAddress = useSelector(hasBadAddressSelector);
-  const profileContactsEnabled = useSelector(selectProfileContactsToggle);
+  const profileContactsPageEnabled = useSelector(
+    selectProfileContactsPageToggle,
+  );
 
   useEffect(() => {
     document.title = `Profile | Veterans Affairs`;
@@ -153,7 +155,7 @@ export const Hub = () => {
           </div>
         </HubCard>
 
-        {profileContactsEnabled && (
+        {profileContactsPageEnabled && (
           <HubCard
             heading={PROFILE_PATH_NAMES.CONTACTS}
             content="Review your next of kin and emergency contact details for accuracy and correctness."
