+++ src/applications/personalization/profile/constants.js
@@ -1,6 +1,6 @@
 // all the active feature toggles for the profile app with a default value of false
 export const PROFILE_TOGGLES = {
-  profileContacts: false,
+  profileContactsPage: false,
   profileShowPronounsAndSexualOrientation: false,
   profileHideDirectDepositCompAndPen: false,
   profileShowPaymentsNotificationSetting: false,
