+++ src/applications/personalization/profile/selectors.js
@@ -161,10 +161,10 @@ export const selectHideDirectDepositCompAndPen = state =>
 
 export const selectIsBlocked = state => cnpDirectDepositIsBlocked(state);
 
-export const selectProfileContactsToggle = state =>
-  toggleValues(state)?.[FEATURE_FLAG_NAMES.profileContacts];
+export const selectProfileContactsPageToggle = state =>
+  toggleValues(state)?.[FEATURE_FLAG_NAMES.profileContactsPage];
 
-export const selectProfileContacts = state => state?.profileContacts || false;
+export const selectProfileContacts = state => state?.profileContacts || {};
 
 export const selectEmergencyContact = state => {
   const contacts = selectProfileContacts(state).data || [];
