+++ src/platform/utilities/feature-toggles/featureFlagNames.js
@@ -123,7 +123,8 @@ export default Object.freeze({
   pdfWarningBanner: 'pdf_warning_banner',
   pensionFormEnabled: 'pension_form_enabled',
   preEntryCovid19Screener: 'pre_entry_covid19_screener',
-  profileContacts: 'profile_contacts',
+  profileContactsV0: 'profile_contacts_v0',
+  profileContactsPage: 'profile_contacts_page',
   profileDoNotRequireInternationalZipCode: 'profile_do_not_require_international_zip_code',
   profileHideDirectDepositCompAndPen: 'profile_hide_direct_deposit_comp_and_pen',
   profileLighthouseDirectDeposit: 'profile_lighthouse_direct_deposit',
